Splits initializers
