+++ app/maintainers/demo_maintainer.rb
@@ -1,4 +1,4 @@
-# Perform maintenance tasks
+# Performs maintenance tasks
 class DemoMaintainer
   include MaintainerHelper
 end
