+++ config/initializers/hidata.rb
@@ -1,0 +1,1 @@
+HIDATA_LINK = Hidata.link
