+++ config/initializers/mini_i18n.rb
@@ -1,0 +1,6 @@
+MiniI18n.configure do |config|
+  config.load_translations('./config/locales/*')
+  config.default_locale = :en
+  config.available_locales = [:en]
+  config.fallbacks = false
+end
