+++ config/initializers/oj.rb
@@ -1,0 +1,1 @@
+Oj.default_options = { mode: :compat, allow_blank: true }
